+++ circonus/__init__.py
@@ -1,5 +1,5 @@
 __title__ = "circonus"
-__version__ = "0.0.11"
+__version__ = "0.0.12"
 
 from logging import NullHandler