+++ circonus/client.py
@@ -16,9 +16,11 @@ import json
 from circonus.annotation import Annotation
 from circonus.collectd.cpu import get_cpu_graph_data
 from circonus.collectd.df import get_df_graph_data
+from circonus.collectd.graph import get_collectd_graph_data
 from circonus.collectd.memory import get_memory_graph_data
 from circonus.collectd.interface import get_interface_graph_data
 from circonus.tag import get_tags_with, get_telemetry_tag, is_taggable
+from requests import codes as status_codes
 from requests.exceptions import HTTPError
 
 import requests
@@ -69,7 +71,11 @@ def with_common_tags(f):
 
 
 def log_http_error(f):
-    """Decorator for :py:mod:`logging` any :class:`~requests.exceptions.HTTPError` raised by a request."""
+    """Decorator for :py:mod:`logging` any :class:`~requests.exceptions.HTTPError` raised by a request.
+
+    :raises: :class:`requests.exceptions.HTTPError`
+
+    """
     @wraps(f)
     def wrapper(*args, **kwargs):
         try:
@@ -78,6 +84,7 @@ def log_http_error(f):
         except HTTPError as e:
             log.error("%s: %s (%s)", e.response.status_code, e.response.json()["code"],
                       e.response.json()["message"])
+            raise
         return r
     return wrapper
 
@@ -228,20 +235,34 @@ class CirconusClient(object):
         a.create()
         return a
 
+    def get_collectd_check_bundle(self, target):
+        """Get a ``collectd`` check bundle for ``target``.
+
+        :param str target: The target to filter the check bundle.
+        :rtype: :py:class:`dict`
+
+        The *first* ``collectd`` check bundle for ``target`` will be returned.  A given ``target`` should not have
+        more than one ``collectd`` check bundle at any given moment.
+
+        """
+        r = self.get("check_bundle", {"f_target": target, "f_type": "collectd"})
+        check_bundle = r.json()[0]
+        return check_bundle
+
     def create_collectd_cpu_graph(self, target):
         """Create a CPU graph from a ``collectd`` check bundle for ``target``.
 
         :param str target: The target of the check bundle to filter for.
-        :rtype: :class:`requests.Response`
+        :rtype: :class:`requests.Response` or :py:const:`None`
 
         ``target`` is used to filter ``collectd`` check bundles.
 
+        :py:const:`None` is returned if no data to create the graph could be found for ``target``.
+
         """
-        r = self.get("check_bundle", {"f_target": target, "f_type": "collectd"})
-        r.raise_for_status()
-        check_bundle = r.json()[0]
+        check_bundle = self.get_collectd_check_bundle(target)
         data = get_cpu_graph_data(check_bundle)
-        return self.create("graph", data)
+        return self.create("graph", data) if data else None
 
     def create_collectd_memory_graph(self, target):
         """Create a memory graph from a ``collectd`` check bundle for ``target``.
@@ -254,16 +275,15 @@ class CirconusClient(object):
         :py:const:`None` is returned if no data to create the graph could be found for ``target``.
 
         """
-        r = self.get("check_bundle", {"f_target": target, "f_type": "collectd"})
-        r.raise_for_status()
-        check_bundle = r.json()[0]
+        check_bundle = self.get_collectd_check_bundle(target)
         data = get_memory_graph_data(check_bundle)
         return self.create("graph", data) if data else None
 
-    def create_collectd_interface_graph(self, target):
+    def create_collectd_interface_graph(self, target, interface_name="eth0"):
         """Create an interface graph from a ``collectd`` check bundle for ``target``.
 
         :param str target: The target of the check bundle to filter for.
+        :param str interface_name: (optional) The interface name, e.g., "eth0".
         :rtype: :class:`requests.Response` or :py:const:`None`
 
         ``target`` is used to filter ``collectd`` check bundles.
@@ -271,10 +291,8 @@ class CirconusClient(object):
         :py:const:`None` is returned if no data to create the graph could be found for ``target``.
 
         """
-        r = self.get("check_bundle", {"f_target": target, "f_type": "collectd"})
-        r.raise_for_status()
-        check_bundle = r.json()[0]
-        data = get_interface_graph_data(check_bundle)
+        check_bundle = self.get_collectd_check_bundle(target)
+        data = get_interface_graph_data(check_bundle, interface_name)
         return self.create("graph", data) if data else None
 
     def create_collectd_df_graph(self, target, mount_dir):
@@ -289,8 +307,46 @@ class CirconusClient(object):
         :py:const:`None` is returned if no data to create the graph could be found for ``target``.
 
         """
-        r = self.get("check_bundle", {"f_target": target, "f_type": "collectd"})
-        r.raise_for_status()
-        check_bundle = r.json()[0]
+        check_bundle = self.get_collectd_check_bundle(target)
         data = get_df_graph_data(check_bundle, mount_dir)
         return self.create("graph", data) if data else None
+
+    def create_collectd_graphs(self, target, interface_names=None, mount_dirs=None):
+        """Create several graphs from a ``collectd`` check bundle.
+
+        :param str target: The target of the check bundle to filter for.
+        :param list interface_name: (optional) The interface names to create ``interface`` graphs for, e.g., ``["eth0"]``.
+        :param list mount_dirs: (optional) The mount directories to create ``df`` graphs for, e.g., ``["/root", "/mnt"]``.
+        :rtype: (:py:class:`bool`, :py:class:`list`)
+
+        ``target`` is used to filter ``collectd`` check bundles.
+
+        Only the first ``collectd`` check bundle will be used as no ``target`` should have more than a single
+        ``collectd`` check bundle at once.
+
+        ``mount_dirs`` should be a :py:class:`list` of directories where devices are mounted.  ``df`` graphs will be
+        created for each.
+
+        ``interface_names`` should be a :py:class:`list` of network interface device names.  ``interface`` graphs will
+        be created for each.
+
+        Return :py:const:`True` if all ``collectd`` graphs were created successfully, :py:const:`False` otherwise,
+        along with a :py:class:`list` of :class:`requests.Response` instances for requests made.
+
+        """
+        if mount_dirs is None:
+            mount_dirs = ["root"]
+
+        if interface_names is None:
+            interface_names = ["eth0"]
+
+        responses = []
+        try:
+            check_bundle = self.get_collectd_check_bundle(target)
+            graph_data = get_collectd_graph_data(check_bundle, interface_names, mount_dirs)
+            for d in graph_data:
+                responses.append(self.create("graph", d))
+        except HTTPError as e:
+            log.error("collectd graphs could not be created: %s", e)
+
+        return responses and all([status_codes.OK == r.status_code for r in responses]), responses